+++ diesel/src/macros/mod.rs
@@ -1169,6 +1169,8 @@ mod query_id;
 mod static_cond;
 #[macro_use]
 mod ops;
+#[macro_use]
+mod tuples;
 
 #[cfg(test)]
 mod tests {