+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## Unreleased
+
+### Added
+
+* Added support for PG tuples. See [`sql_types::Record`][record-1-3-0] for details.
+
+[record-1-3-0]: http://docs.diesel.rs/diesel/pg/types/sql_types/struct.Record.html
+
 ## [1.2.2] - 2018-04-12
 
 ### Changed